refactor: move the resources component into the resources folder
