+++ carrerait-roadmap/src/app/recursos/page.tsx
@@ -1,4 +1,4 @@
-import Resources from "@/components/Resources";
+import Resources from "@/components/Resources/Resources";
 export default function ResourcesPage() {
   return <Resources />;
 }
